+++ bluebottle/clients/postgresql_backend/base.py
@@ -1,8 +1,17 @@
 from tenant_schemas.postgresql_backend import base
 
+from bluebottle.clients import properties
+
 
 class DatabaseWrapper(base.DatabaseWrapper):
-    pass
+    def set_tenant(self, tenant, *args, **kwargs):
+        """
+        Main API method to current database schema,
+        but it does not actually modify the db connection.
+        """
+        super(DatabaseWrapper, self).set_tenant(tenant, *args, **kwargs)
+
+        properties.set_tenant(tenant)
 
 
 DatabaseError = base.DatabaseError
