+++ bluebottle/clients/utils.py
@@ -36,7 +36,6 @@ class LocalTenant(object):
     def __enter__(self):
         if self.tenant:
             connection.set_tenant(self.tenant)
-            properties.set_tenant(self.tenant)
             ContentType.objects.clear_cache()
 
     def __exit__(self, type, value, traceback):
@@ -47,7 +46,7 @@ class LocalTenant(object):
             except AttributeError:
                 logger.info("Attempted to clear missing tenant properties.")
         elif self.previous_tenant:
-            properties.set_tenant(self.previous_tenant)
+            connection.set_tenant(self.tenant)
 
 
 def tenant_url():
